requirements.txt
anthropic>=0.40.0
PyQt6>=6.6.0
Pillow>=10.0.0
python-dotenv>=1.0.0
mss>=9.0.0
dbus-python>=1.3.2; sys_platform=="linux"
